Critical Thinking and Writing by Donna Haraway
Donna J. Haraway (b. September 6, 1944) is an American Professor Emerita in the History of Consciousness Department and Feminist Studies Department at the University of California, Santa Cruz, United States. She is a prominent scholar in the field of science and technology studies, described in the early 1990s as a "feminist and postmodernist".